Implements LogOut functionality and page refresh.
