+++ src/navBar/NavBar.css
@@ -53,7 +53,7 @@
   font-weight: bold;
   border: 1px solid;
   border-radius: 4px;
-  min-width: 160px;
+  min-width: 190px;
   margin: 20px auto;
 }
